Add Patrick's sawtooth-sdk types needed for signing
- Access example: `sawtooth.signing.Signer`
