--- src/types/sawtooth-sdk/signing.d.ts
@@ -1,3 +1,0 @@
-// Type definitions for Sawooth SDK Signing library
-// Project: https://www.hyperledger.org/projects/sawtooth, https://github.com/hyperledger/sawtooth-core
-// Definitions by: Patrick Erichsen <https://github.com/Patrick-Erichsen>
